extract MacroCallTransformingVisitor
